BAU: Only run retag step if required
Instead of always running the retag step in the deployment pipelines,
and then conditionally doing nothing based on an environment variable,
we can instead conditionally include the retag step in the pipeline
based on the value of the `retag_image_on_success` variable.

Having a step called `pull-image-retag-and-push` that sometimes does
nothing is confusing when viewing the pipeline in the AWS Console.