BAU: stop using timestamp for final_snapshot_identifier
This setting is only used when destroying the RDS cluster, so using
timestamp() causes unnecessary drift on every plan.

Further to that, it could actually introduce confusion - the timestamp
on the final snapshot would have no relation to when it was actually
destroyed (if it was destroyed some time after the plan was applied).