BAU: Add `track_latest` to pipeline visualiser
Also, use `task.arn` to reference ECS task definitions, rather than
`.id` or `family:revision`. `arn` returns the full ARN, including the
revision, which is cleaner than constructing it manually.

`track_latest` works the same way with the full arn as it does with
`family:revision`, so this change is safe.