BAU: Update Mailchimp sync task with forms-admin
Previously, the ECS task definition for the Mailchimp sync task was
not being updated when a new container image was deployed for
forms-admin. This meant that the terraform had to have been applied
separately after each deployment in order to update the task definition
with the new image.

Now, the CodePipeline for deploying the forms-admin container
includes a new CodeBuild action which updates the Mailchimp sync task
definition with the new image URI after each deployment. The EventBridge
rule which triggers the Mailchimp sync task uses the task family name
instead of a specific revision ARN, so it will always use the latest
revision of the task definition.

Also, the log stream prefix for the Mailchimp sync task has been updated
to use the same prefix as the forms-admin application, for consistency.