Add keywords to pipeline Slack notifications
Try to visually highlight details about Slack notifications using the
"keywords" formatting that AWS ChatBot provides [[1]]. Currently it can
be hard to quickly pick out what environment and app an alert is related
to.

This does lead to a bit of repetition in the content, but I think that's
fine, and if we don't like it we can keep iterating.

[1]: https://docs.aws.amazon.com/chatbot/latest/adminguide/custom-notifs.html