Add action to pipeline failure Slack notifications
It is useful to be able to see at a glance what action in a pipeline
a failure occurred in. This commit changes the Slack notifications for
pipeline failures to trigger on the CodePipeline action events, so we
can get the name of the action from the EventBridge event and add it to
the message template.