Send RSpec output to EventBridge whenever end to end tests fail
Currently to see why the end to end tests have failed for a deploy a
developer has to connect to the VPN and sign in to AWS. It would be nice
instead if we could see the output of the test run in a Slack
notification.

Normally the output is only in CloudWatch logs, this commit sets things up
so that additionally the output of RSpec is sent to EventBridge.

We arrange things so that only the RSpec test report output is sent,
ignoring other log messages such as debugging/tracing, to keep noise to
a minimum. We do this by telling RSpec to output both to stdout and a
file, and then send the contents of that file to EventBridge, along with
some useful metadata.

The structure of the "CodeBuild run-e2e-tests RSpec output" detail type
is chosen to mimic AWS's "CodeBuild Build State Change" detail type.