Remove the WAF for the ALB
We already have a WAF at the Cloudfront level. We needed two when we couldn’t
guarantee that all requests arriving at the ALB were coming from Cloudfront. But
we have configured CloudFront to add a custom HTTP header to requests that it
sends to the ALB, and the ALB to only forward requests that contain that header.