Apply rule to all objects in the bucket
The default behaviour for an empty `filter` is to apply the rule to all objects.
Terraform recommends adding this argument (even though it's optional):

"The filter argument, while Optional, is required if the rule configuration
block does not contain a prefix and you intend to override the default behavior
of setting the rule to filter objects with the empty string prefix (""). Since
prefix is deprecated by Amazon S3 and will be removed in the next major version
of the Terraform AWS Provider, we recommend users specify filter."

We tried adding this before but we were encountering an error:

╷
│ Warning: Invalid Attribute Combination
│
│   with module.forms_runner.aws_s3_bucket_lifecycle_configuration.this,
│   on ../../../modules/forms-runner/file-upload.tf line 16, in resource "aws_s3_bucket_lifecycle_configuration" "this":
│   16: resource "aws_s3_bucket_lifecycle_configuration" "this" {
│
│ No attribute specified when one (and only one) of
│ [rule[0].filter,rule[0].prefix] is required
│
│ This will be an error in a future version of the provider
│
│ (and one more similar warning elsewhere)

It looks like they have fixed it