docker-compose: Add default service name to Compose file
Rather than letting Docker Compose use `local` as the project name (from
the directory name), which isn't very descriptive, set the default
project name to `forms`.

If for some reason this isn't desirable it can be overridden with a
command line flag or environment variable [[1]].

Also rename services so that they are not prefixed `forms-forms-`.

[1]: https://docs.docker.com/compose/how-tos/project-name/