docker-compose: Fix warnings from Bundler
We were getting `/home/ruby` is not writable errors from Bundler,
because we made the container root readonly in PR #1569. This commit
fixes that error by mounting a read-write volume over the top for the
bundler folder at `~/.bundle`, similar to how we fixed a similar error
from npm in commit ef37c03ed6.