Manage AWS ChatBot configurations in Terraform
As the aws provider now supports AWS ChatBot resources, we can manage
the previously manually created ChatBot configurations in Terraform.

Rather than hard-coding ARNs for the ChatBot SNS topics in various
places, we have a well-known module that defines and exposes the names
and ARNs of these topics so they can be referenced from other
deployments.