BAU: Add more instructions for dependabot tf PRs
GitHub protects us from recursion by not running workflows on commits
made by GitHub bots. This could be solved later by using a custom bot
user with a PAT, but for now we can just add instructions to the PR
body to tell maintainers to amend the PR to get the checks to run.

Also, things can get a bit messy if dependabot tries to update
dependencies on a PR where the lockfiles have already been updated.
So add a warning to tell maintainers to recreate the PR if that
happens.