BAU: always run report job if no failure or cancel
Rather than having two identically named jobs for pre-commit, instead
have a single job with a final report job that always runs if there
are no failures or cancellations.

Also, apply the `if` condition to the terraform-ci test job.

This means we will no longer see "skipped" jobs in the GitHub Actions UI
for these workflows, which is cleaner.