BAU: add a workflow to request team reviews on PRs
We could use `CODEOWNERS` for this, but it requests the review from
the team as a whole, rather than individuals, which means that if one
person reviews it, the team is no longer requested for review.

Also, some IDEs will show when a user is requested for review but not
when a team is requested for review, so this makes it easier to see when
a review has been requested of you.