BAU: Improve tflint cache key
Without the architecture and OS in the cache key, different
architectures (e.g. arm64 and amd64) will share the same cache.

If a PR is opened when the job is configured to run on one architecture,
then later the job configuration is changed to run on a different
architecture, the cache will be restored from the previous architecture's
cache, which will likely be invalid for the new architecture.