Add environment name to environment variables for end to end test runs
We want to make the configuration of the end to end tests more similar
to the configuration of the apps. The first step is making the end to
end tests aware of the environment it is running against. There isn't an
equivalent of `RAILS_ENV` for the end to end tests, but we can set the
`forms_env` setting and use that in a similar way.