BAU: Cache ruby deps for CodeBuild in S3
Rather than downloading and installing all ruby dependencies anew for
each CodeBuild run, cache them in S3 and restore from there.

The cache is always restored owned by `root:root`, so we need to `chown`
the vendor directory back to the `ruby` user before running any commands
that need to access the dependencies. This means we now perform the
`install` phase commands as `root`, and only switch to the `ruby` user
for the `build` phase, where we actually run the tests.

This makes us less reliant on external services (rubygems.org), and
ensures that we always have the same versions of gems for each unique
`Gemfile.lock` state.